Feat: tier component add #112
