+++ frontend/src/components/Profile/mmr/index.js
@@ -1,1 +1,2 @@
-export {default as MMRBoard} from "./mmrBoard.js";
+export { default as MMRBoard } from "./mmrBoard.js";
+export { default as Tier } from "./tier.js";
